Hardcoded function works!!!

+++ src/table.rs
@@ -94,7 +94,7 @@ impl Table {
             Kind::Operator => {
                 self.set_operator(entry, Operator::from_string(raw_value).unwrap().get_pos())
             }
-            Kind::Null => self.set_null(entry),
+            Kind::Null => self.set_null(entry, raw_value != "null"),
             Kind::Function => {
                 /*let (pos, arguments) = raw_value.split_at(raw_value.find(CHAR_SEP_NAME).unwrap());
 
@@ -195,8 +195,12 @@ impl Table {
         }
     }
 
-    pub fn set_null(&mut self, entry: &str) {
-        self.remove_entry(entry);
+    pub fn set_null(&mut self, entry: &str, delete: bool) {
+        if delete {
+            self.remove_entry(entry);
+        } else {
+            self.set(entry, 0, Kind::Null);
+        }
     }
 
     pub fn set_function(&mut self, entry: &str, value: Function) {
